+++ CHANGELOG.md
@@ -8,7 +8,7 @@ This is not a comprehensive list of changes but rather a hand-curated collection
 
 v4.4.1
 ======
-- Update `report.py` to allow a generic list of colors
+- Update `report.py` to set specific colors for plotted trajectories
 - Made `Component::getSocketNames` a `const` member method (previously: non-const)
 
 v4.4